backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.3.0
psutil>=5.9.0
python-dotenv>=1.0.0
redis>=5.0.0
celery>=5.3.0
boto3>=1.34.0

# Monitoring and observability
sentry-sdk[fastapi]>=1.40.0
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client>=0.19.0

# HuggingFace integration
datasets>=2.14.0
huggingface_hub>=0.17.0

# Roboflow integration
roboflow>=1.1.0

# AI/ML models and processing
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0

# Computer Vision models
ultralytics>=8.0.0  # YOLO
segment-anything>=1.0  # SAM2 (or alternative)
depth-anything-v2>=0.1.0  # Depth estimation
clip-by-openai>=1.0  # CLIP embeddings

# Image processing
scikit-image>=0.21.0
albumentations>=1.3.0